Separate TF1 and TF2 tests.
For easier naming.
